Bunch of refactoring
