+++ app/server/utils/metrics/filters.py
@@ -1,0 +1,134 @@
+from server.models.credit_transfer import CreditTransfer
+from server.models.transfer_account import TransferAccount
+from server.utils.metrics import metrics_cache
+from server.models.custom_attribute_user_storage import CustomAttributeUserStorage
+
+from server.utils.transfer_enums import TransferTypeEnum, TransferSubTypeEnum, TransferStatusEnum
+from flask import g
+from server.models.user import User
+from sqlalchemy.sql import func, text
+from sqlalchemy import case, or_
+from server import db, red, bt
+
+def apply_filters(query, filters, query_table):
+
+    if filters is None:
+        return query
+
+    user_join_attribute, account_join_attribute = determine_join_conditions(query_table)
+
+    for filter_table_name, _filts in filters.items():
+        if filter_table_name == TransferAccount.__tablename__:
+                # only join transfer account if table is not transfer account
+                if query_table.__tablename__ == filter_table_name: 
+                    query = apply_single_column_filter(query, _filts, TransferAccount, None, None) 
+                else:
+                    query = apply_single_column_filter(query, _filts, TransferAccount, account_join_attribute, None)
+
+        elif filter_table_name == User.__tablename__:
+                # only join user account if table is not user 
+                if query_table.__tablename__ == filter_table_name:
+                    query = apply_single_column_filter(query, _filts, User, None, None)
+                else:
+                    query = apply_single_column_filter(query, _filts, User, None, user_join_attribute)
+
+        elif filter_table_name == CustomAttributeUserStorage.__tablename__ and user_join_attribute is not None:
+            query = apply_ca_filters(query, _filts, user_join_attribute)
+    return query
+
+def determine_join_conditions(table):
+    if table == CreditTransfer:
+        return CreditTransfer.sender_user_id, CreditTransfer.sender_transfer_account_id
+    if table == User:
+        return User.id, None
+
+def apply_single_column_filter(query, filters, target_table, account_join_attribute=None, user_join_attribute=None):
+    if target_table.__tablename__ == TransferAccount.__tablename__ and account_join_attribute is not None:
+        query = query.join(TransferAccount, TransferAccount.id == account_join_attribute)
+    elif target_table.__tablename__ == User.__tablename__ and user_join_attribute is not None:
+        query = query.join(User, User.id == user_join_attribute)
+
+    for batches in filters:
+        to_batch = []
+        for _filt in batches:
+            column = _filt[0]
+            comparator = _filt[1]
+            val = _filt[2]
+
+            if comparator == 'EQ':
+                val = val if isinstance(val, list) else [val]
+                to_batch.append(getattr(target_table, column).in_(val))
+            elif comparator == 'GT':
+                to_batch.append(getattr(target_table, column) > val)
+            elif comparator == "LT":
+                to_batch.append(getattr(target_table, column) < val)
+        query = query.filter(or_(*to_batch))
+    return query
+
+def apply_ca_filters(query, filters, user_join_condition):
+
+    # get all custom attributes and create pivot table
+    new_cs = [CustomAttributeUserStorage.user_id]
+    for value in db.session.query(CustomAttributeUserStorage.name).distinct():
+        value = value[0]
+        new_cs.append(
+             func.max(case(
+                [(CustomAttributeUserStorage.name == value, CustomAttributeUserStorage.value)],
+                else_=None
+            )).label(value)
+        )
+
+    # join pivot table of custom attributes
+    pivot = db.session.query(*new_cs).group_by(CustomAttributeUserStorage.user_id).subquery()
+    query = query.outerjoin(pivot, user_join_condition == pivot.c.user_id)
+    
+    for batches in filters:
+        to_batch = []
+        for _filt in batches:
+            column = _filt[0]
+            comparator = _filt[1]
+            val = _filt[2]
+
+            if comparator == 'EQ':
+                val = val if isinstance(val, list) else [val]
+                val = [f'\"{element}\"' for element in val] # needs ot be in form '"{item}"' for json string match
+                to_batch.append(pivot.c[column].in_(val))
+            elif comparator == 'GT':
+               to_batch.append(pivot.c[column] > val)
+            elif comparator == "LT":
+                to_batch.append(pivot.c[column] < val)
+        query = query.filter(or_(*to_batch))
+
+    return query
+
+
+disbursement_filters = [
+    CreditTransfer.transfer_status == TransferStatusEnum.COMPLETE,
+    CreditTransfer.transfer_type == TransferTypeEnum.PAYMENT,
+    CreditTransfer.transfer_subtype == TransferSubTypeEnum.DISBURSEMENT
+]
+
+standard_payment_filters = [
+    CreditTransfer.transfer_status == TransferStatusEnum.COMPLETE,
+    CreditTransfer.transfer_type == TransferTypeEnum.PAYMENT,
+    CreditTransfer.transfer_subtype == TransferSubTypeEnum.STANDARD
+]
+
+exchanged_filters = [
+    CreditTransfer.transfer_status == TransferStatusEnum.COMPLETE,
+    CreditTransfer.transfer_type == TransferTypeEnum.EXCHANGE,
+    #CreditTransfer.token == g.active_organisation.token
+]
+
+beneficiary_filters = [User.has_beneficiary_role == True]
+vendor_filters = [User.has_vendor_role == True]
+
+exhaused_balance_filters = [
+    CreditTransfer.transfer_type == TransferTypeEnum.PAYMENT,
+    TransferAccount._balance_wei == 0
+]
+
+transfer_use_filters = [
+    *standard_payment_filters,
+    CreditTransfer.transfer_use.isnot(None),
+]
